pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "devworld"
version = "0.0.0"
description = "Demo for the talk about AI-agents at DevWorld 2025"
authors = [
    "Alex Shershebnev <alexs@forgood.ai>",
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"

anthropic = "^0.47.0"
mcp = "^1.3.0"
mcp-server-fetch = "^2025.1.17"
notebook = "^7.4.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
